Update 585666b9-adbc-5d04-c480-58e55ea7fb9d.md
Original syntax used in example code would cause a compile error in VBA, paretheses should only be used to pass arguments when there is a return value being assigned to something. Alternatively, a space could be used between the function name and the first parentheses - however this is effectively just grouping the argument(s) and isn't standard practice.